requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
bcrypt>=3.2.0,<4.0
passlib[bcrypt]
python-jose
sqlalchemy[asyncio]
asyncpg
openai
python-dotenv
pydub
alembic